requirements.txt
blinker==1.9.0
certifi==2024.12.14
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
mysql-connector-python==9.2.0
netifaces==0.11.0
numpy==2.0.1
oauthlib==3.2.0
packaging==24.2
pandas==2.2.2
pluggy==1.5.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==8.3.4
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
ranger-fm==1.9.3
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.37
systemd-python==234
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
Werkzeug==3.1.3
zipp==1.0.0
